HSO: minor updates

+++ examples/CarControlSystem/application/src/carcontrol/device/Speed.java
@@ -3,6 +3,6 @@ package carcontrol.device;
 public interface Speed {
 	
 	public int getSpeed();
-	public void speed (double value);
+	public void speed (int value);
 
 }
